requirements.txt
Django
djangorestframework
django_compressor
django-filter
